rest/requirements.txt
asgiref==3.5.2
Django
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
djoser
django-filter
django-allauth
django.utils.six
